Move knownAddresses to global
