+++ code/client/src/state/modules/wallet/actions.js
@@ -27,11 +27,6 @@ const setRelayerSecret = createAction('SET_RELAYER_SECRET')
 const setNetwork = createAction('SET_NETWORK')
 const setProvider = createAction('SET_PROVIDER')
 
-// Set the wallet address that is known by the user, that is the address has been entered in the system at least once.
-// E.g. transaction destination or recovery address.
-const setKnownAddress = createAction('SET_KNOWN_ADDRESS')
-const deleteKnownAddress = createAction('DELETE_KNOWN_ADDRESS')
-
 const bindDomain = createAction('BIND_DOMAIN')
 
 // User has acknowledged that they has been notified to save the wallet address for a created wallet.
@@ -56,8 +51,6 @@ export default {
   setRelayerSecret,
   setNetwork,
   setProvider,
-  setKnownAddress,
-  deleteKnownAddress,
   bindDomain,
 
   userAcknowledgedToSaveAddress,
